prediction considers all task input to be at least 1 byte to avoid multiplication/division by 0

+++ hadoop-tools/hadoop-posum/src/main/java/org/apache/hadoop/tools/posum/common/records/dataentity/JobProfile.java
@@ -30,7 +30,7 @@ public interface JobProfile extends GeneralDataEntity {
 
     void setFinishTime(Long finishTime);
 
-    Integer getDuration();
+    Long getDuration();
 
     void setTotalMapTasks(Integer totalMapTasks);
